V2.07 Mark up.  Fix 64bit build warnings.
git-svn-id: https://svn.code.sf.net/p/veroroute/code/trunk@339 d5f823ff-b8c7-416f-ab5d-d69718ecf442

+++ src/ConnectionMatrix.h
@@ -46,7 +46,7 @@ public:
 	{
 		// Make j-k connection and enforce transitivity
 
-		typedef std::pair<unsigned int, unsigned int> CONNECTION;
+		typedef std::pair<size_t, size_t> CONNECTION;
 		std::list<CONNECTION> list;		// Helper for updating the connection matrix
 		list.push_back( CONNECTION(j,k) );
 		while ( !list.empty() )
@@ -60,7 +60,7 @@ public:
 			{
 				m_pp[a][b] = m_pp[b][a] = true;	// Make a-b connection ...
 				m_cost -= 2;					// Update cost
-				for (unsigned int c = 0; c < m_N; c++)	// Update 1st-order transitive relations
+				for (size_t c = 0; c < m_N; c++)	// Update 1st-order transitive relations
 				{
 					if ( m_pp[a][c] )
 					{
